+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Added missing type for `cds.context.model`
 - Added missing type for `cds.context.model`
 - Added missing type for `req.query.elements`
+- Made constructors for query parts (`SELECT`, `UPDATE`, `DELETE`, ...) private, as they should only be accessed statically
 
 ### Added
 - `cds.app` typed as express.js application
@@ -18,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `cds.utils.colors` types
 - The CQL methods `.where` and `.having` now suggest property names for certain overloads.
 - `Service.before/on/after(event, target...)` now accept also an array of typer-generated classes in the `target` parameter
+- `localized` variants to `SELECT`
 
 ### Changed
 - Most `cds.requires` entries are now optionals.